+++ test/create_mock_dbs.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 ./mysql/create_mysql_mock.sh
-#./mysql/create_mssql_mock.sh
+#./mssql/create_mssql_mock.sh