+++ test/mssql/03-tests-insert_batch.js
@@ -48,37 +48,37 @@ describe('insert_batch()', () => {
     it('should only accept an array as the second parameter', () => {
         qb.reset_query();
 
-        expect(() => qb.insert_batch('galaxies',test_data), 'array of objects provided').to.not.throw(Error);
-        expect(() => qb.insert_batch('galaxies',[]), 'empty array provided').to.not.throw(Error);
+        expect(() => qb.insert_batch('galaxies', test_data), 'array of objects provided').to.not.throw(Error);
+        expect(() => qb.insert_batch('galaxies', []), 'empty array provided').to.not.throw(Error);
 
-        expect(() => qb.insert_batch('galaxies',[{},{}]), 'array of empty objects provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',[test_data,test_data]), 'array of arrays provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',{}), 'empty object provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',''), 'empty string provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',null), 'null provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',undefined), 'undefined provided').to.throw(Error);
         expect(() => qb.insert_batch('galaxies'), 'nothing provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',3), 'integer provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',3.5), 'float provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',true), 'true provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',Infinity), 'Infinity provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',[1,2]), 'array of numbers provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',[Date, /foobar/, null]), 'array of non-standard objects provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',['abc',2,{foo:'bar'}]), 'array of mixed values provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',/foobar/), 'regex provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',NaN), 'NaN provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',false), 'false provided').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies','  '), 'string full of spaces provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', [{},{}]), 'array of empty objects provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', [test_data,test_data]), 'array of arrays provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', {}), 'empty object provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', ''), 'empty string provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', null), 'null provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', undefined), 'undefined provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', 3), 'integer provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', 3.5), 'float provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', true), 'true provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', Infinity), 'Infinity provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', [1,2]), 'array of numbers provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', [Date, /foobar/, null]), 'array of non-standard objects provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', ['abc',2,{foo:'bar'}]), 'array of mixed values provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', /foobar/), 'regex provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', NaN), 'NaN provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', false), 'false provided').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', '  '), 'string full of spaces provided').to.throw(Error);
     });
     it('should allow for an empty data parameter', () => {
         qb.reset_query();
-        const sql = qb.insert_batch('galaxies',[]);
+        const sql = qb.insert_batch('galaxies', []);
         sql.should.eql("INSERT INTO [galaxies] () VALUES ()");
     });
     it('should utilize pre-existing tables set in from_array', () => {
         qb.reset_query();
         qb.from('galaxies');
-        const sql = qb.insert_batch(null,[]);
+        const sql = qb.insert_batch(null, []);
         sql.should.eql("INSERT INTO [galaxies] () VALUES ()");
     });
     it('should fail if any invalid values are passed into one of the data objects in the dataset', () => {
@@ -88,17 +88,30 @@ describe('insert_batch()', () => {
         const arr = [1,2,3];
         const obj = {foo: 'bar'};
 
-        expect(() => qb.insert_batch('galaxies',[{id: func}]), 'function in data').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',[{id: regex}]), 'regex in data').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',[{id: Infinity}]), 'Infinity in data').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',[{id: undefined}]), 'undefined in data').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',[{id: NaN}]), 'NaN in data').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',[{id: arr}]), 'array in data').to.throw(Error);
-        expect(() => qb.insert_batch('galaxies',[{id: obj}]), 'object in data').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', [{id: func}]), 'function in data').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', [{id: regex}]), 'regex in data').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', [{id: Infinity}]), 'Infinity in data').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', [{id: undefined}]), 'undefined in data').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', [{id: NaN}]), 'NaN in data').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', [{id: arr}]), 'array in data').to.throw(Error);
+        expect(() => qb.insert_batch('galaxies', [{id: obj}]), 'object in data').to.throw(Error);
     });
-    it('should support insert ignore statements', () => {
+    it('should not support insert ignore statements', () => {
         qb.reset_query();
-        const sql = qb.insert_batch('galaxies', test_data, true, 'ON DUPLICATE KEY UPDATE last_update = NOW()');
-        sql.should.eql("INSERT IGNORE INTO [galaxies] ([id], [name], [type]) VALUES (3, 'Milky Way', 'spiral'), (4, 'Andromeda', 'spiral') ON DUPLICATE KEY UPDATE last_update = NOW()");
+        expect(() => qb.insert_batch('galaxies', test_data, true)).to.throw(Error);
     });
+    it('should not support insert statement suffixes', () => {
+        qb.reset_query();
+        expect(() => qb.insert_batch('galaxies', test_data, false, 'ON DUPLICATE KEY UPDATE')).to.throw(Error);
+    });
+    // it('should include the OUTPUT directive when the return() method is called in the chain', () => {
+    //     qb.reset_query();
+    //     const sql = qb.returning('id').insert_batch('galaxies', test_data);
+    //     sql.should.eql("INSERT INTO [galaxies] ([id], [name], [type]) OUTPUT INSERTED.[id] VALUES (3, 'Milky Way', 'spiral')");
+    // });
+    // it('should include the OUTPUT directive with multiple outputs when the return() method is called in the chain with an array', () => {
+    //     qb.reset_query();
+    //     const sql = qb.returning(['id', 'name']).insert_batch('galaxies', test_data);
+    //     sql.should.eql("INSERT INTO [galaxies] ([name], [type]) OUTPUT INSERTED.[id], INSERTED.[name] VALUES ('Milky Way', 'spiral')");
+    // });
 });